refactor(styling): a ton more styling cleanup : - restructured more clearly ui into top, left, and bottom toolbars - cleanup & update pass for css - further spliting up of css into a per component basis
