+++ src/components/EntityInfos/index.js
@@ -4,29 +4,7 @@ import { combineLatestObj, exists } from '../../utils/obsUtils'
 import view from './view'
 import intent from './intent'
 
-// //////
-import ColorPicker from '../widgets/ColorPicker'
-
-export function colorPickerWrapper (state$, DOM) {
-  console.log('making colorPicker')
-  const props$ = // just({color:"#FF00FF"})
-  state$.map(function (state) {
-    let {meta, transforms} = state
-
-    if (!meta || !transforms) {
-      return undefined
-    }
-    if (transforms.length > 0) transforms = transforms[0]
-    if (meta.length > 0) meta = meta[0]
-
-    return {color: meta.color}
-  })
-
-  return ColorPicker({DOM, props$})
-}
-
-// //////
-
+////////
 function model (props$, actions) {
   let comments$ = props$.pluck('comments').filter(exists).startWith(undefined)
   let meta$ = props$.pluck('meta').filter(exists).startWith(undefined)
