+++ src/components/EntityInfos/style.css
@@ -1,0 +1,136 @@
+/*entity infos*/
+
+.toolbarBottom{
+  font-size: 13px;
+  position: absolute;
+  bottom: 0px;
+  left: 50px;
+  background: white;
+  width: 100%;
+  padding: 10px;
+}
+
+.nameInput {
+  width: 200px !important;
+  max-width: none !important;
+  font-weight: 700;
+}
+
+.toolbarBottom input[type=text]{
+  box-sizing:border-box;
+  -moz-box-sizing:border-box;
+
+  border: none;
+  border-width: 0px;
+
+
+  max-width: 150px;
+  border: none;
+  line-height: 0.8em;
+  font-size: 0.8em;
+  font-family: inherit;
+
+  text-overflow: ellipsis;
+}
+
+.toolbarBottom input[type=number]{
+  box-sizing:border-box;
+  -moz-box-sizing:border-box;
+
+  border: none;
+  border-width: 0px;
+
+  width:10em;
+  max-width: 50px;
+  border: none;
+  line-height: 0.8em;
+  font-size: 0.8em;
+  font-family: inherit;
+}
+
+/*.tooolbarBottom */input[type=color]{
+  width: 24px;
+  margin: 0;
+  border: 0px;
+  background: none;
+  padding: 0px;
+  height: 20px;
+}
+
+input[type=color]:focus{
+  outline: none;
+}
+
+.toolbarBottom input[type=text]:focus{
+  outline: none;
+}
+
+.toolbarBottom input[type=number]:focus{
+  outline: none;
+}
+
+.entityInfos .icon{
+  width: 16px;
+  height: 16px;
+  top: 4px;
+  position: relative;
+  padding: 0;
+}
+
+.entityInfos .icon path{
+  fill:#989898;
+}
+
+.entityInfos .icon:hover path{
+  background:none;
+  color:none;
+  fill:#989898;
+}
+
+
+.entityInfos .transformInput{
+  padding: 7 0 0 13px;
+}
+
+
+.color{
+  width: 24px;
+  margin: 0;
+  border: 0px;
+  padding: 0px;
+  height: 20px;
+  display: inline;
+  outline: none;
+
+      position: static;
+    float: left;
+    bottom: 0px;
+    left: 0px;
+}
+
+.colorPicker{
+
+}
+
+.colorPicker .body{
+  width: 150px;
+  height: 90px;
+  position: absolute;
+  bottom: 35px;
+  border: 1px solid #D8D8D8;
+  padding: 5px;
+}
+
+.colorPicker .body .mainGradient{
+  height: 70px;
+  width: 100%;
+  margin: 2px;
+}
+
+.colorPicker .body .smallGradient{
+  height: 20px;
+  width: 100%;
+  margin: 2px;
+  background: linear-gradient(to right, red, orange, yellow, green, blue, indigo, violet);
+
+}
