+++ src/components/main/leftToolBar.css
@@ -1,0 +1,70 @@
+
+#leftToolbar{
+  /*background: #fff; #f5f5f5;*/
+  padding: 0;/*5 0 0 10; */
+
+  position: absolute;
+  top: 10px;
+  left:10px;
+  width: 180px;
+  height: 90%;
+}
+
+#leftToolbar h1{
+  display: inline-block;
+  padding: 4px;
+  margin: 0px;
+  margin-left: 5px;
+  font-size: 1.3em;
+  font-weight: 500;
+
+  min-width: 200px;
+}
+
+/* text entry*/
+#leftToolbar input[type=text]{
+  color:#60c4f8;
+  outline: none;
+  font-size: 1.1em;
+  font-weight: 500;
+
+  min-width: 10em;
+  width:100%;
+  box-sizing:border-box;
+  -moz-box-sizing:border-box;
+
+  border: none;
+  border-width: 0px;
+}
+
+#leftToolbar input[type=text]:focus{
+  outline: none;
+}
+
+#leftToolbar .history{
+    padding-left: 5px;
+  padding-right: 5px;
+}
+
+#leftToolbar .social{
+  padding-left: 5px;
+  padding-right: 5px;
+}
+
+#leftToolbar section{
+  display: inline-flex;;
+  /* padding-left: 15px;
+  padding-right: 25px;
+      width: 50px;*/
+          flex-direction: column;
+}
+
+#leftToolbar button {
+  width:50px;
+  height:50px;
+  margin: 2px;
+}
+#leftToolbar button .icon{
+  width:20px;
+  height:20px;
+}
