+++ src/components/main/notifications.css
@@ -1,0 +1,7 @@
+.notifications{
+  color: #A5A5A5;
+  font-size: 0.7em;
+  margin-top: 10px;
+  vertical-align: text-top;
+  text-align: center;
+}
