+++ src/components/widgets/Icon/style.css
@@ -1,0 +1,19 @@
+/*ICONS*/
+.icon{
+  /*width: 24px;
+  height: 24px;
+  padding: 4px;*/
+  width: 20px;
+  height: 20px;
+  padding: 8px;
+  /*stroke: #60C4F8;
+  stroke-width: 2;*/
+}
+
+.icon path{
+  fill:#dddddd;
+}
+
+.icon:hover path{
+  fill:#60C4F8;
+}
