+++ src/components/widgets/Menu/index.js
@@ -3,7 +3,7 @@ import { html } from 'snabbdom-jsx'
 import Class from 'classnames'
 import assign from 'fast.js/object/assign'// faster object.assign
 
-import style from './menu.css'
+import style from './style.css'
 //require('./menu.css')
 
 function getToolTip (tooltip, toggleCondition) {
