+++ src/components/widgets/ProgressBar/index.js
@@ -1,7 +1,8 @@
-
 import { html } from 'snabbdom-jsx'
 import { combineLatestObj } from '../../../utils/obsUtils'
 import { exists } from '../../../utils/utils'
+require('./style.css')
+
 
 function model (props$, actions) {
   const progress$ = props$.pluck('progress')
